'fixed classname typo and move icons to file'
