+++ src/components/Header/header.scss
@@ -36,7 +36,7 @@ $navbarHeight: 40px;
     align-items: center;
   }
 
-  &.network {
+  .network {
     margin-right: 5px;
     display: flex;
     flex-direction: column;
